+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.11.0
  - feat: add ability to customize `NesContainer`s pixel sizes.
+ - feat: add ability to customize `NesButton`s pixel sizes.
 
 # 0.10.0
  - feat: add `NesIcons.user`